+++ backend/src/baserow/contrib/integrations/local_baserow/receivers.py
@@ -7,5 +7,5 @@ from baserow.core.cache import global_cache, local_cache
 @receiver(table_schema_changed)
 def invalidate_table_cache(sender, table_id, **kwargs):
     # Invalidate local cache when the table schema is updated
-    global_cache.invalidate(f"table_{table_id}__service_schema")
+    global_cache.invalidate(invalidate_key=f"table_{table_id}__service_invalidate_key")
     local_cache.delete(f"integration_service_{table_id}_table_model")